feat(observer): add message queue with /enqueue endpoint for sequential turns (fixes NV-7742)
Adds per-session message queuing to the thalamus observer DO:

- New /enqueue endpoint: reserves a queue slot, returns active/queued
- message_queue SQLite table: stores RequestParams (no credentials)
- Queue state machine: idle → active on first message, queued messages
  drain via queue-ready webhook when the current turn completes
- toolResults bypass the queue (resolve requires-action immediately)
- TTL sweep on startup purges stale entries (>10 min)
- Refactored into separate modules: session-observer, types, parsers,
  validation, worker
